fix snr function from detected spots
